renamed all instances of env vars
