+++ .github/workflows/updateMembers.yml
@@ -9,8 +9,8 @@ jobs:
     steps:
       - name: Call Supabase Edge Function
         env:
-          SUPABASE_URL: ${{ secrets.VITE_SUPABASE_URL }}
+          SUPABASE_URL: ${{ secrets.SUPABASE_URL }}
           SUPABASE_SERVICE_ROLE_KEY: ${{ secrets.SUPABASE_SERVICE_ROLE_KEY }}
         run: |
           curl -X POST "$SUPABASE_URL/functions/v1/updateMembers" \
-               -H "Authorization: Bearer $VITE_SUPABASE_SERVICE_ROLE_KEY"
+               -H "Authorization: Bearer $SUPABASE_SERVICE_ROLE_KEY"
