+++ supabase/functions/updateMembers/index.ts
@@ -5,12 +5,12 @@ import { config } from "https://deno.land/x/dotenv/mod.ts";
 console.log("Current working directory:", Deno.cwd());
 
 const env = config();
-const SUPABASE_URL = env.VITE_SUPABASE_URL;
-const SUPABASE_SERVICE_ROLE_KEY = env.VITE_SUPABASE_SERVICE_ROLE_KEY;
-const CONGRESS_API_KEY = env.VITE_CONGRESS_API_KEY;
+const SUPABASE_URL = env.SUPABASE_URL;
+const SUPABASE_SERVICE_ROLE_KEY = env.SUPABASE_SERVICE_ROLE_KEY;
+const CONGRESS_API_KEY = env.CONGRESS_API_KEY;
 
 if (!SUPABASE_URL || !SUPABASE_SERVICE_ROLE_KEY || !CONGRESS_API_KEY) {
-  throw new Error("Missing required environment variables: VITE_SUPABASE_URL, VITE_SUPABASE_SERVICE_ROLE_KEY, or VITE_CONGRESS_API_KEY");
+  throw new Error("Missing required environment variables: SUPABASE_URL, SUPABASE_SERVICE_ROLE_KEY, or CONGRESS_API_KEY");
 }
 
 const supabase = createClient(SUPABASE_URL, SUPABASE_SERVICE_ROLE_KEY);
