+++ vite.config.ts
@@ -1,6 +1,7 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 import tailwindcss from '@tailwindcss/vite'
+import envVars from "../LearnCongress/src/utils/loadEnv"
 
 
 // https://vite.dev/config/
@@ -10,6 +11,9 @@ export default defineConfig({
     tailwindcss(),
     
   ],
+  define: {
+    "process.env": envVars,
+  },
   server: {
     watch: {
       usePolling: true
